requirements.txt
streamlit
PyPDF2
PyMuPDF
Pillow
requests
reportlab
openai